refactor: rework crowd file names for clarity
